thirdWheel: just use move() instead of correct()
In the previous commit, a correct() function was created that was an
exact clone of move(), except that it moved slower. Why didn't I just
change move() to be able to take power and tolerance as input? Because I
am a dumb-dumb. Anyway, correct() was removed and move() was refactored.